requirements.txt
cloudinary==1.32.0
fastapi==0.94.1
gunicorn==20.1.0
numpy==1.24.2
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas==1.5.3
Pillow==9.4.0
pydantic==1.10.6
pytest==7.2.2
python-dotenv==1.0.0
python-multipart==0.0.6
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
uuid==1.30
uvicorn==0.21.1
jinja2
opencv-python-headless
